backend/requirements.txt
django==3.2
djangorestframework==3.13
django-filter==21.1
django_extra_fields==3.0.2
djoser==2.1.0
psycopg2-binary==2.8.6
gunicorn==20.1.0
python-dotenv==0.20.0